fix: Updating core components and adding outline-config.
