+++ packages/outline-config/postcss.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  plugins: [
+    require('postcss-import'),
+    require('tailwindcss'),
+    require('postcss-nested'),
+    require('postcss-custom-properties'),
+    require('autoprefixer'),
+    require('postcss-discard-comments'),
+  ],
+};
