+++ packages/outline-core/package.json
@@ -26,6 +26,7 @@
     "package": "yarn publish"
   },
   "dependencies": {
+    "@phase2/outline-config": "^0.0.1",
     "lit": "^2.0.2"
   },
   "publishConfig": {
