+++ packages/outline-demonstrations/index.ts
@@ -1,3 +1,6 @@
 export { OutlineSectionLink } from './src/outline-section-link/outline-section-link';
 export { OutlineSwatch } from './src/outline-swatch/outline-swatch';
-export { OutlineSwatchSet } from './src/outline-swatch-set/outline-swatch-set';
+export {
+  OutlineSwatchSet,
+  swatchSets,
+} from './src/outline-swatch-set/outline-swatch-set';
