+++ packages/outline-storybook/package.json
@@ -10,5 +10,12 @@
     "type": "git",
     "url": "https://github.com/phase2/outline.git",
     "directory": "packages/outline-storybook"
+  },
+  "dependencies": {
+    "@phase2/outline-demonstrations": "^0.0.1",
+    "lit": "^2.0.2"
+  },
+  "publishConfig": {
+    "access": "public"
   }
 }
