+++ packages/outline-storybook/stories/components/demonstration/outline-swatch-set.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Canvas } from '@storybook/addon-docs';
 import { html } from 'lit';
-import { OutlineSwatchSet, swatchSets } from './outline-swatch-set';
+import { OutlineSwatchSet, swatchSets } from '@phase2/outline-demonstrations';
 
 <Meta
   title="Utilities/Swatch Set"
