+++ packages/outline-storybook/stories/components/demonstration/outline-swatch.stories.mdx
@@ -1,7 +1,6 @@
 import { Meta, Story, Canvas } from '@storybook/addon-docs';
 import { html } from 'lit';
-import { OutlineSwatch } from './outline-swatch';
-import { swatchSets } from '../outline-swatch-set/outline-swatch-set';
+import { OutlineSwatch, swatchSets } from '@phase2/outline-demonstrations';
 
 <Meta
   title="Utilities/Swatch"
