+++ packages/outline-storybook/stories/tokens/colors.stories.mdx
@@ -1,7 +1,6 @@
 import { html, unsafeHTML } from 'lit';
 import { Meta, Canvas, Story, Preview } from '@storybook/addon-docs';
-import tailwindThemeConfig from '../../../resolved-tailwind-config';
-import '@phase2/outline-swatch-set'
+import '@phase2/outline-demonstrations'
 
 <Meta
   title="Design Tokens/Colors/Solid"
