+++ packages/outline-templates/default/svgo.config.js
@@ -1,30 +1,4 @@
+const defaultConfig = require(`@phase2/outline-config/svgo.config`);
 module.exports = {
-  multipass: true,
-  js2svg: {
-    indent: 2, // string with spaces or number of spaces. 4 by default
-    pretty: true, // boolean, false by default
-  },
-  plugins: [
-    {
-      name: 'preset-default',
-      params: {
-        overrides: {
-          cleanupNumericValues: {
-            floatPrecision: 2,
-          },
-          // disable plugins
-          anotherBuiltinPlugin: false,
-          addAttributesToSVGElement: false,
-          addClassesToSVGElement: false,
-          removeElementsByAttr: false,
-          removeAttrs: false,
-          removeRasterImages: false,
-          removeUnusedNS: false,
-          removeViewBox: false,
-          removeXMLNS: false,
-          sortAttrs: false,
-        },
-      },
-    },
-  ],
+  ...defaultConfig,
 };
