+++ postcss.config.js
@@ -1,10 +1,4 @@
+const defaultConfig = require(`@phase2/outline-config/postcss.config`);
 module.exports = {
-  plugins: [
-    require('postcss-import'),
-    require('tailwindcss')('./tailwind.config.js'),
-    require('postcss-nested'),
-    require('postcss-custom-properties'),
-    require('autoprefixer'),
-    require('postcss-discard-comments'),
-  ],
+  ...defaultConfig,
 };
